build: target .NET Framework 4.8 and Magick.NET 14.17.1
Retarget both projects from .NET Framework 4.0 Client to 4.8 and move
from Magick.NET 7.16.1 to 14.17.1 plus Magick.NET.Core. Visual Studio
regenerated the settings, resources and dataset designer files.

Port to the 14.x API: generic pixel collections and images, uint sizes,
ImageMagick.Drawing namespace. MagickWrapper converts System.Drawing
colors with MagickColor.FromRgba, since the 16 bit constructor would
take 0-255 values unscaled. The water texture resize is dropped: the
sub percent resize was a no-op in 7.x but shrinks the tile to 1x1 in 14.

Renders of zones 163, 169 and 171 match the 7.16.1 output (mean
absolute difference 0.1 to 0.4 of 255).

+++ MapCreator/Classes/ZoneConfiguration.cs
@@ -206,7 +206,7 @@ namespace MapCreator
         public ImageMagick.MagickImage GetWaterMap()
         {
             ImageMagick.MagickImage watermap = new ImageMagick.MagickImage(MpkWrapper.GetFileBytesFromMpk(this.m_datMpk, "water.pcx"));
-            watermap.Resize(this.m_targetMapSize, this.m_targetMapSize);
+            watermap.Resize((uint)(this.m_targetMapSize), (uint)(this.m_targetMapSize));
 
             return watermap;
         }
